+++ pymobiledevice/instruments.py
@@ -30,7 +30,7 @@ from ctypes import Structure,c_byte,c_uint16,c_uint32
 from socket import inet_ntoa,htons,inet_ntop,AF_INET6
 
 from . import bplist
-from . import struct2 as ct
+from .util import struct2 as ct
 from .exceptions import MuxError, ServiceError
 from .lockdown import LockdownClient
