+++ pymobiledevice/util/struct2.py
